+++ CHANGELOG.md
@@ -1,11 +1,20 @@
+## 0.2.1-dev.1
+
+ADD:
+
+- config service.
+- add formatTime in date_util.
+
 ## 0.2.0-dev.2
 
 FIX
+
 - function exports.
 
 ## 0.2.0-dev.1
 
 ADD:
+
 - image crop screen/services.
 - picker services.
 - cache services.